Add signs array to its own file.
