+++ src/Components/AllSigns/AllSigns.js
@@ -1,81 +1,8 @@
-import aries from '../../images/aries.png';
-import taurus from '../../images/taurus.png';
-import gemini from '../../images/gemini.png';
-import cancer from '../../images/cancer.png';
-import leo from '../../images/leo.png';
-import virgo from '../../images/virgo.png';
-import libra from '../../images/libra.png';
-import scorpio from '../../images/scorpio.png';
-import sagittarius from '../../images/sagittarius.png';
-import capricorn from '../../images/capricorn.png';
-import aquarius from '../../images/aquarius.png';
-import pisces from '../../images/pisces.png';
+import './AllSigns.css'
+import allSigns from '../../allSigns';
 import SignCard from '../SignCard/SignCard';
 
 const AllSigns = () => {
-  const allSigns = [
-    {
-      sign: 'aries',
-      image: aries,
-      id: 1
-    },
-    {
-      sign: 'taurus',
-      image: taurus,
-      id: 2
-    },
-    {
-      sign: 'gemini',
-      image: gemini,
-      id: 3
-    },
-    {
-      sign: 'cancer',
-      image: cancer,
-      id: 4
-    },
-    {
-      sign: 'leo',
-      image: leo,
-      id: 5
-    },
-    {
-      sign: 'virgo',
-      image: virgo,
-      id: 6
-    },
-    {
-      sign: 'libra',
-      image: libra,
-      id: 7
-    },
-    {
-      sign: 'scorpio',
-      image: scorpio,
-      id: 8
-    },
-    {
-      sign: 'sagittarius',
-      image: sagittarius,
-      id: 9
-    },
-    {
-      sign: 'capricorn',
-      image: capricorn,
-      id: 10
-    },
-    {
-      sign: 'aquarius',
-      image: aquarius,
-      id: 11
-    },
-    {
-      sign: 'pisces',
-      image: pisces,
-      id: 12
-    }
-  ];
-
   const signDisplay = allSigns.map(sign => {
     return (
       <SignCard
