+++ src/Components/SignCard/SignCard.css
@@ -1,3 +1,7 @@
 .sign-image {
   height: 150px;
 }
+
+.sign-card {
+  width: 200px;
+}
